refactor: course-header.ts -> course-route.header.ts
