+++ src/routes/_authenticated/courses/$courseSlug/attempts.tsx
@@ -1,5 +1,5 @@
 import { requireCourseRole } from '@/course/course.guards';
-import { createCourseSectionBreadcrumb } from '@/course/course-header';
+import { createCourseSectionBreadcrumb } from '@/course/course-route.header';
 import { createFileRoute } from '@tanstack/react-router';
 
 export const Route = createFileRoute(
