+++ src/routes/_authenticated/courses/$courseSlug/repositories/$studentUsername/commits/route.tsx
@@ -1,4 +1,4 @@
-import { getRepositoryCommitsBreadcrumb } from '@/course/course-header';
+import { getRepositoryCommitsBreadcrumb } from '@/course/course-route.header';
 import { Outlet, createFileRoute } from '@tanstack/react-router';
 
 export const Route = createFileRoute(
